Deleted data files, migrated to ZIO 2
