--- zioflow_rocksdb/CURRENT
@@ -1,1 +1,0 @@
-MANIFEST-000006
